feat: add useHistory hook
